Create Niftiprofile app
The app was created using the command `django-admin startapp
niftiprofile`. The app should hold the logic for displaying the user
profiles that visitors see when they visit the website.

Creating this app is necessary for separating the view from the logic.

An app called `niftisearch` may be needed later on for holding the logic
of searching these profiles.