Create home, user, and search apps.
The applications were created using the `startapp` command, for example
`django-admin startapp home`.

The name `profile` would have been more suitable than `user` but could
not be used because it conflicted with a Python module.

The `home` app is needed for routing to the `about`, `help`, `terms`, and
any other static pages that do not access the database.

The `search` app is needed for retrieving database information, and
displaying users that match the query. The search app will send the
information to the `user` app, which will display it.

The `user` app is needed for routing to the user profiles, and will
contain the logic needed to display user information retrieved from the
database.

The `user` app replaces the `niftiprofile` app because it has a more
concise name.