Separate header and footer from base template.
The `base.html` template file was getting cluttered, so I decided to
turn the header and footer into individual files.

I also replaced the `text-center` class from the `body` tag with the
`text-align: center` style because it was forcing every text element to
be centered, and could not be overriden. The `text-align` property has
the same functionality, but can be overriden.