Add search results section.
The search results section is the part of the website where the profiles
of each user are loaded. The results show a picture and a description of
each profile.

The search results sections is the interface thru which visitors will
search the website users. Visitors should be able to click on user
profiles and be redirected to the profiles.